+++ CHANGES.rst
@@ -21,6 +21,10 @@
 
 - Fix the logging of some environment variables RelStorage uses.
 
+- If there is a read conflict error, PostgreSQL no longer holds any
+  database locks while the error is raised and the transaction is
+  rolled back in Python. Previously, shared locks could be held during
+  this process, preventing other transactions from moving forward.
 
 3.4.5 (2021-04-23)
 ==================